Mostly works
All the key pieces are in place. But maybe this should be on button
press? Also need to figure out dynamic entry/switching of words for
papers when that button is chosen.

Also maybe there should be a zoom in feature to go in towards the
points.

Also I think I could/should have precomputed euclidean distances for up
to 100 IDs (our max). The euclidean distance calculation is slow!